Have mouse events pass a struct to the callback
The struct is more flexible to add changes useful for the callback than
the original CForge::Mouse object. For now this is only used for a
cursor position adjusted with the position where the widget was
initially clicked, so that this adjusted position can be directly used
for the widget's setPosition() [which seems to work better on my Linux-
based system]. This struct could also be expanded to hold the status of
shift, ctrl and alt to allow for more advanced UI features, should the
need arise.