Implement a FormWidget with callback
Implements a FormWidget that can group multiple inputs together and adds
an "Apply" button that triggers the callback. For that to work, another
Widget was added, the LabelWidget. It groups an input with its
accompanying text label and provides a more abstract getValue function
based on pointers instead of explicit values. Thus, the callback can use
a single map for all its settings instead of handling each value type
individually. The callback code only needs to cast the pointers back to
their original/expected type.