Fix WidgetBackground rendering
The bottom value of the vertex data was positive, thus actually at the
top. Therefore the polygon would end up inverted and removed by the
backface culling.

Also, the z coordinate had to be set or text would not render on the
background. Might need require some changes in the future.