Add a WindowWidget and methods to create them
Adds a new WindowWidget that can be minimized and moved around.
A function to easily create one containing a FormWidget was also
added to the GUI class. It returns a reference to the form for the
calling code to add its options to it. Cleanup should be handled by
the GUI.