Add text input and update font loading
Adds a text input widget with an override-able method to validate input.
This could be used to implement a floating point number input, convert
the entered text to some other type or implement restrictions on what
may be entered.

Font loading and character map texture generation was changed to better
support the arbitrary user input. The changes make it easier to change
what characters should be loaded and lock the data structure after it's
done. This way, user input outside of the loaded range won't produce
broken looking text but instead fall back to the proper placeholder
glyph.