Add multi-line text capabilities to TextWidget
Adds multi line text capabilities. To use them, pass TextWidget::setText
a string containing line breaks ('\n'). You can also set the alignment
to left, right, or centered.

Also implemented but not yet exposed through the TextWidget API are
maximum widths. The method that breaks the string into individual lines,
FontFace::computeStringWidthMultiline, does respect it (and even has an
option for a maximum number of allowed text lines). But the TextWidget
does not yet have a way to make use of this.